% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_collect_pageview}
\alias{ga_collect_pageview}
\title{Send pageviews to Google Analytics}
\usage{
ga_collect_pageview(page_url = NULL, page = NULL, title = NULL,
  hostname = galog$settings$hostname, user_id = NULL)
}
\arguments{
\item{page_url}{a character string with the url of the page being visited}

\item{page}{a character string with the page which was visited}

\item{title}{Optional. A character string with the title of the page which was visited}

\item{hostname}{a character string with the hostname. Defaults TEMP if not set}

\item{user_id}{Optional. A user ID to assign to your visit. A random client_id is generated alongside the user_id.}
}
\value{
a list with the generated client_id and user_id
}
\description{
Send pageviews to Google Analytics.
If someone visits a page, use this function to send the page and title of the page which is visited so that you can
easily see how users are using your application. \cr

Pageviews can be viewed in the Google Analytics > Behaviour tab or in the Real-Time part of Google Analytics.
}
\examples{

ga_collect_pageview(page = "/home")
user <- ga_collect_pageview(page_url="www.foo.com/mypage/")
user <- ga_collect_pageview(page = "/home", title = "Homepage", hostname = "www.foo.com")
user <- ga_collect_pageview(page_url="www.foo.com/mypage.hml")
ga_collect_pageview(page = "/home", user_id=user)
user <- ga_collect_pageview(page=session$clientData$url_pathname, hostname=session$clientData$url_hostname)
}
