% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_delete_settings}
\alias{ga_delete_settings}
\title{Delete user settings file}
\usage{
ga_delete_settings(path = NULL)
}
\arguments{
\item{path}{path to a file to delete}
}
\description{
Delete user settings file
}
\examples{
ga_delete_settings(path="/location/to/your/settings/file.json")
}
