% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_get_settings}
\alias{ga_get_settings}
\title{Save settings specified by user to file}
\usage{
ga_get_settings(path)
}
\arguments{
\item{path}{location of the file to use}
}
\value{
the settings in a JSON format
}
\description{
Save settings specified by user to file
}
\examples{
ge_get_settings()
}
