% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_initialize}
\alias{ga_initialize}
\title{Initialize tracking}
\usage{
ga_initialize(path = NULL, tracking_id = NULL, user_id = NULL,
  hostname = "TEMP", consent = TRUE)
}
\arguments{
\item{path}{Path to the file where the settings are saved (Optional). If provided, settings are loaded from that location. Further parameters are ignored}

\item{tracking_id}{}

\item{user_id}{}

\item{hostname}{}

\item{consent}{}
}
\description{
Initialize tracking
}
\examples{
ga_initialize("~/galog.json")
ga_initalize(tracking_id="UA-XXXXXX-x,",user_id="Known user",hostname="www.foo.com",concent=TRUE)
}
