% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_save_settings}
\alias{ga_save_settings}
\title{Set user defined settings to file}
\usage{
ga_save_settings(path = NULL, ...)
}
\arguments{
\item{path}{location of file to use}

\item{...}{settings to save. Name-value pairs.}
}
\value{
a helpful message
}
\description{
Set user defined settings to file
}
\examples{
ga_set_settings(tracking_id="UA-XXXXXXXX-x",hostname="www.Foo.com",consent=TRUE)
}
