% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_set_approval}
\alias{ga_set_approval}
\title{Request for approval of the user to send information to Google Analytics}
\usage{
ga_set_approval(message, consent = FALSE)
}
\arguments{
\item{message}{a character string with a message to show to the user before approval is requested. If not given will show a default message.}

\item{consent}{logical indicating to give approval. Defaults to FALSE indicating that no approval is given.}
}
\value{
invisibly a list all general settings used to send data to Google Analytics
}
\description{
Request for approval of the user to send information to Google Analytics.
The approval is requested by setting yes/no in the prompt if consent is set to FALSE and the user is working in interactive mode.\cr
If consent is set to TRUE, the developer is responsible for complying to legislation in your area like the 
General Data Protection Regulation (GDPR) in Europa.
}
\examples{
## Request user input
ga_set_approval(consent = FALSE)
ga_set_approval(consent = FALSE,
                message = "Please approve that we send usage information to Google Analytics")

## Developer sets consent directly assuming that he received approval in another way
ga_set_approval(consent = TRUE)
}
