% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_set_client_id}
\alias{ga_set_client_id}
\title{Set client id}
\usage{
ga_set_client_id(client_id = NULL)
}
\arguments{
\item{client_id}{a character string in UUID format which anonymously and uniquely identifies a particular R user or device across different R sessions.
Defaults to a randomly generated UUID.}
}
\value{
client_id
}
\description{
Defaults to a randomly generated identifier.
}
\details{
You can also set the client_id identifier which anonymously identifies a particular user or device. 
For R users this client_id identifies the same user across different R sessions. 
The value of this field should be a random UUID (version 4) as described in \url{http://www.ietf.org/rfc/rfc4122.txt}
}
\examples{
ga_set_client_id(client_id=uuid::UUIDgenerate())
}
