% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_set_hostname}
\alias{ga_set_hostname}
\title{Set hostname}
\usage{
ga_set_hostname(hostname = "Google.com")
}
\arguments{
\item{host_name}{the hostname as character}
}
\value{
galog object with hostname in
}
\description{
Set hostname
}
\examples{
ga_set_hostname("Magical hostname")
}
