% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_set_tracking_id}
\alias{ga_set_tracking_id}
\title{Provide the Google Analytics tracking ID where all user interactions will be logged to}
\usage{
ga_set_tracking_id(x = "UA-25938715-4")
}
\arguments{
\item{x}{a character string with the Google Analytics tracking ID}
}
\value{
invisibly a list all general settings used to send data to Google Analytics
}
\description{
The Google Analytics tracking ID looks something like UA-XXXXX-Y. For example UA-25938715-4.
Get your own tracking ID at the Google Analytics website.
All collected user interactions will be logged with that tracking ID.
}
\examples{
ga_set_tracking_id("UA-25938715-4")
}
