% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_initialize}
\alias{ga_initialize}
\title{Initialize tracking}
\usage{
ga_initialize(path = NULL, tracking_id = NULL, consent = TRUE,
  hostname = NULL, message = NULL)
}
\arguments{
\item{path}{Path to the file where the settings are saved (Optional). If provided, settings are loaded from that location. Further parameters are ignored}

\item{tracking_id}{Google Analytics property ID (UA-XXXXXXXX-x)}

\item{consent}{TRUE/FALSE. Do you agree to provide data to Google analytics.}

\item{hostname}{Optinal. Hostname}

\item{message}{Optional. Custom message to display upon agreement of terms.}
}
\description{
Initialize tracking
}
\examples{
ga_initialize("~/galog.json")
ga_initalize(tracking_id="UA-XXXXXXXX-x,",hostname="www.foo.com",concent=TRUE)
}
