% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_save_settings}
\alias{ga_save_settings}
\title{Set user defined settings to file}
\usage{
ga_save_settings(path = NULL, ...)
}
\arguments{
\item{path}{location of file to use}

\item{...}{settings to save. Must be a list with name-value pair (e.g. tracking_id="XXXXXXXX-x")}
}
\value{
a helpful message
}
\description{
Set user defined settings to file
}
\examples{
ga_set_settings(user_id="User1",client_id="111111-x",hostname="Google.com",consent=TRUE)
}
