% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_set_approval_message}
\alias{ga_set_approval_message}
\title{Set the aproval message for the user}
\usage{
ga_set_approval_message(message = NULL)
}
\arguments{
\item{x}{sepcify a custom message; otherwise use default message}
}
\value{
the message
}
\description{
Set the aproval message for the user
}
