% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_set_url}
\alias{ga_set_url}
\title{Create galog url}
\usage{
ga_set_url()
}
\value{
galog object
}
\description{
Creates galog$url and ready to send to Google
}
\details{
v=1              // Version. 
 &tid=UA-XXXXX-Y  // Tracking ID / Property ID.
 &cid=555         // This anonymously identifies a particular user, device,
 or browser instance 
 &uid=555         // Known identifier for a user provided
 by the site owner/tracking library user 
 &ds=GAlogger        // Data source: set to GAlogger
 
 Google protocol guidelines:
 
https://developers.google.com/analytics/devguides/collection/protocol/v1/devguide
https://developers.google.com/analytics/devguides/collection/protocol/v1/parameters
}
