% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ga_set_user_id}
\alias{ga_set_user_id}
\title{Provide the identifier which will be used to identify a visitor/user}
\usage{
ga_set_user_id(user_id = NULL)
}
\arguments{
\item{user_id}{a character string with the visitor/user known to you. Defaults to a randomly generated UUID.}
}
\value{
invisibly a list all general settings used to send data to Google Analytics
}
\description{
Set the identifier of a visitor as it is known by you. 
The user_id identifier is the identifier of the visitor/user as it is know by you.
By default for every new R session, a new client_id is generated.
}
\examples{
ga_set_user_id()
ga_set_user_id("root")
ga_set_user_id("team-datascientists")
ga_set_user_id("shiny-server")

x <- sprintf("\%s-\%s", Sys.getpid(), tolower(Sys.getenv("USERNAME", unset = "default")))
x
ga_set_user_id(x)
}
